+++ src/main.rs
@@ -1,15 +1,8 @@
 use clap::Parser;
 use color_eyre::eyre::Result;
-use neverest::{cli::Cli, printer::StdoutPrinter};
+use neverest::cli::Cli;
 
 #[tokio::main]
 async fn main() -> Result<()> {
-    neverest::tracing::install()?;
-
-    let cli = Cli::parse();
-    let mut printer = StdoutPrinter::new(cli.output, cli.color);
-
-    cli.command
-        .execute(&mut printer, cli.config_paths.as_ref())
-        .await
+    Cli::parse().execute().await
 }